JVxeTable出现循环依赖问题
